feat: move relation  e2e tests to dedicated file 2
